0-basic_async_syntax.py

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -6,6 +6,6 @@ import random
 
 async def wait_random(max_delay: int = 10) -> int:
     """function to await a number and return it"""
-    await_number: int = random.uniform(0, max_delay)
+    await_number: float = random.uniform(0.0, float(max_delay))
     await asyncio.sleep(await_number)
     return await_number
